Move server app to the separate file
